+++ tests/test_util.py
@@ -80,140 +80,3 @@ def test_long_path_zip_file():
                 in result
             )
 
-
-def test_format_like_automatic1111_sampler_unmatched():
-    class Bounds:
-        width = 256
-        height = 256
-
-    class Params:
-        seed = 42
-        bounds = Bounds()
-        metadata = {
-            "prompt": "Test",
-            "negative_prompt": "",
-            "sampler": "UnknownSampler",
-            "checkpoint": "unknown.ckpt",
-            "loras": [],
-        }
-
-    result = util._format_like_automatic1111(Params())
-    assert "Sampler: UnknownSampler" in result
-    assert "Steps: Unknown" in result
-    assert "CFG scale: Unknown" in result
-
-
-def test_format_like_automatic1111_loras_dict_and_tuple():
-    class Bounds:
-        width = 128
-        height = 128
-
-    class Params:
-        seed = 1
-        bounds = Bounds()
-        metadata = {
-            "prompt": "Prompt",
-            "negative_prompt": "",
-            "sampler": "Euler - euler_a (10 / 5.0)",
-            "checkpoint": "loramodel.ckpt",
-            "loras": [{"name": "lora1", "weight": 0.7}, ("lora2", 0.5), ["lora3", 0.9]],
-        }
-
-    result = util._format_like_automatic1111(Params())
-    assert "<lora:lora1:0.7>" in result
-    assert "<lora:lora2:0.5>" in result
-    assert "<lora:lora3:0.9>" in result
-
-
-def test_format_like_automatic1111_strength_none_and_one():
-    class Bounds:
-        width = 64
-        height = 64
-
-    class ParamsNone:
-        seed = 0
-        bounds = Bounds()
-        metadata = {
-            "prompt": "Prompt",
-            "negative_prompt": "",
-            "sampler": "Euler - euler_a (5 / 2.0)",
-            "checkpoint": "model.ckpt",
-            "strength": None,
-            "loras": [],
-        }
-
-    class ParamsOne:
-        seed = 0
-        bounds = Bounds()
-        metadata = {
-            "prompt": "Prompt",
-            "negative_prompt": "",
-            "sampler": "Euler - euler_a (5 / 2.0)",
-            "checkpoint": "model.ckpt",
-            "strength": 1.0,
-            "loras": [],
-        }
-
-    result_none = util._format_like_automatic1111(ParamsNone())
-    result_one = util._format_like_automatic1111(ParamsOne())
-    assert "Denoising strength" not in result_none
-    assert "Denoising strength" not in result_one
-
-
-def test_format_like_automatic1111_missing_metadata_fields():
-    class Bounds:
-        width = 100
-        height = 200
-
-    class Params:
-        seed = 999
-        bounds = Bounds()
-        metadata = {}
-
-    result = util._format_like_automatic1111(Params())
-    assert "Prompt: " in result
-    assert "Negative prompt: " in result
-    assert "Steps: Unknown" in result
-    assert "Sampler: " in result
-    assert "CFG scale: Unknown" in result
-    assert "Seed: 999" in result
-    assert "Size: 100x200" in result
-    assert "Model: Unknown" in result
-
-
-def test_add_png_itxt_valid(tmp_path):
-    # Create a minimal valid PNG file
-    png_header = b"\x89PNG\r\n\x1a\n"
-    ihdr_chunk = (
-        b"\x00\x00\x00\rIHDR" + b"\x00\x00\x00\x01" + b"\x00\x00\x00\x01" + b"\x08\x02\x00\x00\x00"
-    )
-    ihdr_crc = struct.pack(
-        ">I",
-        zlib.crc32(b"IHDR" + b"\x00\x00\x00\x01\x00\x00\x00\x01\x08\x02\x00\x00\x00") & 0xFFFFFFFF,
-    )
-    iend_chunk = b"\x00\x00\x00\x00IEND" + struct.pack(">I", zlib.crc32(b"IEND") & 0xFFFFFFFF)
-    png_data = png_header + ihdr_chunk + ihdr_crc + iend_chunk
-
-    file_path = tmp_path / "test.png"
-    file_path.write_bytes(png_data)
-
-    util._add_png_itxt(file_path, "testkey", "testvalue")
-
-    # Check that the file still starts with PNG header
-    data = file_path.read_bytes()
-    assert data.startswith(png_header)
-    # Check that iTXt chunk is present
-    assert b"iTXt" in data
-    assert b"testkey" in data
-    assert b"testvalue" in data
-
-
-def test_add_png_itxt_invalid(tmp_path):
-    # Not a PNG file
-    file_path = tmp_path / "not_png.txt"
-    file_path.write_text("not a png")
-    try:
-        util._add_png_itxt(file_path, "key", "value")
-        assert False, "Should have raised ValueError"
-    except ValueError as e:
-        assert "Not a valid PNG file" in str(e)
